# fee-relayer/swagger/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Fee Relayer
  description: Community Points Fee Relayer Service

consumes:
  - application/json
produces:
  - application/json

definitions:
  tx_input:
    required:
      - amount
      - blknum
      - currency
      - oindex
      - owner
      - txindex
    properties:
      amount:
        oneOf:
          - type: object
          - type: integer
          - type: string
      blknum:
        type: integer
      currency:
        type: string
      oindex:
        type: integer
      owner:
        type: string
      txindex:
        type: integer

  tx_output:
    required:
      - amount
      - outputGuard
      - outputType
    properties:
      amount:
        oneOf:
          - type: object
          - type: integer
          - type: string
      outputGuard:
        type: string
      outputType:
        type: integer

  tx:
    required:
      -inputs
      -outputs
    properties:
      inputs:
        type: array
        items:
          $ref: "#/definitions/tx_input"
      outputs:
        type: array
        items:
          $ref: "#/definitions/tx_output"


  create-relayed-tx-body:
    required:
      - utxos
      - amount
      - to
    properties:
      utxos:
        type: array
        items:
          $ref: "#/definitions/tx_input"
      amount:
        oneOf:
          - type: object
          - type: integer
          - type: string
      to:
        type: string

  submit-relayed-tx-body:
    required:
      - tx
      - signatures
    properties:
      tx:
        schema:
          $ref: "#/definitions/tx"
      signatures:
        type: array
        items:
          type: string

  cancel-relayed-tx-body:
    required:
      - tx
    properties:
      tx:
        type: object


paths:
  /create-relayed-tx:
    post:
      description: Creates a transaction
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/create-relayed-tx-body"
      responses:
        200:
          description: Returns the transaction

  /submit-relayed-tx:
    post:
      description: Submits a transaction
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/submit-relayed-tx-body"
      responses:
        200:
          description: Returns the submitted transaction receipt

  /cancel-relayed-tx:
    post:
      description: Cancels a transaction
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/cancel-relayed-tx-body"
      responses:
        200:
          description: Returns true on success